Move guess_scheme() to scrapy.utils.url
